requirements.txt
gym
pygame
stable-baselines3
pyyaml
numpy
scikit-image
